+++ src/Data/Morpheus/Validation/Internal/Utils.hs
@@ -1,4 +1,4 @@
-module Data.Morpheus.Validation.Utils.Utils
+module Data.Morpheus.Validation.Internal.Utils
   ( differKeys
   , existsObjectType
   , lookupType
@@ -6,6 +6,7 @@ module Data.Morpheus.Validation.Utils.Utils
   , lookupField
   , checkNameCollision
   , checkForUnknownKeys
+  , isEqOrStricter
   , VALIDATION_MODE(..)
   ) where
 
@@ -13,7 +14,7 @@ import           Data.List                               ((\\))
 import           Data.Morpheus.Error.Variable            (unknownType)
 import           Data.Morpheus.Types.Internal.Base       (EnhancedKey (..), Key, Position, enhanceKeyWithNull)
 import           Data.Morpheus.Types.Internal.Data       (DataInputType, DataKind (..), DataLeaf (..), DataOutputObject,
-                                                          DataTypeLib (..))
+                                                          DataTypeLib (..), DataTypeWrapper (..))
 import           Data.Morpheus.Types.Internal.Validation (Validation)
 import qualified Data.Set                                as S
 import           Data.Text                               (Text)
@@ -51,10 +52,8 @@ getInputType typeName' lib error' =
             Just (CustomScalar x) -> pure (ScalarKind x)
             Just (LeafEnum x)     -> pure (EnumKind x)
 
-existsObjectType ::
-     Position -> Text -> DataTypeLib -> Validation DataOutputObject
-existsObjectType position' typeName' lib =
-  lookupType error' (object lib) typeName'
+existsObjectType :: Position -> Text -> DataTypeLib -> Validation DataOutputObject
+existsObjectType position' typeName' lib = lookupType error' (object lib) typeName'
   where
     error' = unknownType typeName' position'
 
@@ -67,19 +66,21 @@ removeDuplicates = S.toList . S.fromList
 elementOfKeys :: [Text] -> EnhancedKey -> Bool
 elementOfKeys keys' EnhancedKey {uid = id'} = id' `elem` keys'
 
-checkNameCollision ::
-     [EnhancedKey] -> ([EnhancedKey] -> error) -> Either error [EnhancedKey]
+checkNameCollision :: [EnhancedKey] -> ([EnhancedKey] -> error) -> Either error [EnhancedKey]
 checkNameCollision enhancedKeys errorGenerator =
   case enhancedKeys \\ removeDuplicates enhancedKeys of
     []         -> pure enhancedKeys
     duplicates -> Left $ errorGenerator duplicates
 
-checkForUnknownKeys ::
-     [EnhancedKey]
-  -> [Text]
-  -> ([EnhancedKey] -> error)
-  -> Either error [EnhancedKey]
+checkForUnknownKeys :: [EnhancedKey] -> [Text] -> ([EnhancedKey] -> error) -> Either error [EnhancedKey]
 checkForUnknownKeys enhancedKeys' keys' errorGenerator' =
   case filter (not . elementOfKeys keys') enhancedKeys' of
     []           -> pure enhancedKeys'
     unknownKeys' -> Left $ errorGenerator' unknownKeys'
+
+isEqOrStricter :: [DataTypeWrapper] -> [DataTypeWrapper] -> Bool
+isEqOrStricter [] []                               = True
+isEqOrStricter (NonNullType:xs1) (NonNullType:xs2) = isEqOrStricter xs1 xs2
+isEqOrStricter (NonNullType:xs1) xs2               = isEqOrStricter xs1 xs2
+isEqOrStricter (ListType:xs1) (ListType:xs2)       = isEqOrStricter xs1 xs2
+isEqOrStricter _ _                                 = False